+++ node.gyp
@@ -343,6 +343,7 @@
         'src/module_wrap.cc',
         'src/node.cc',
         'src/node_api.cc',
+        'src/node_binding.cc',
         'src/node_buffer.cc',
         'src/node_config.cc',
         'src/node_constants.cc',
@@ -414,6 +415,7 @@
         'src/node.h',
         'src/node_api.h',
         'src/node_api_types.h',
+        'src/node_binding.h',
         'src/node_buffer.h',
         'src/node_constants.h',
         'src/node_context_data.h',
